Refactor: Continue previous task
