+++ app/src/main/java/com/github/meeplemeet/ui/components/BusinessComponents.kt
@@ -10,6 +10,7 @@ import android.content.Context
 import android.text.format.DateFormat
 import android.widget.Toast
 import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
@@ -24,6 +25,7 @@ import androidx.compose.material.icons.filled.ChevronRight
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Email
+import androidx.compose.material.icons.filled.ExpandMore
 import androidx.compose.material.icons.filled.Language
 import androidx.compose.material.icons.filled.Phone
 import androidx.compose.material.icons.filled.Place
@@ -34,8 +36,12 @@ import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.*
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.ClipboardManager
@@ -544,16 +550,16 @@ fun DayRow(dayName: String, value: String, onEdit: () -> Unit, modifier: Modifie
         Text(
             text = dayName,
             modifier = Modifier.weight(1f).testTag(ShopComponentsTestTags.DAY_ROW_NAME),
-            style = MaterialTheme.typography.bodyLarge)
+            fontSize = Dimensions.TextSize.subtitle)
         Text(
             text = value,
             modifier =
                 Modifier.weight(2f)
                     .padding(end = Dimensions.Spacing.medium)
                     .testTag(ShopComponentsTestTags.DAY_ROW_VALUE),
-            style = MaterialTheme.typography.bodyMedium,
             color = MaterialTheme.colorScheme.onSurfaceVariant,
             textAlign = TextAlign.End,
+            fontSize = Dimensions.TextSize.subtitle
         )
         IconButton(
             onClick = onEdit, modifier = Modifier.testTag(ShopComponentsTestTags.DAY_ROW_EDIT)) {
@@ -1237,4 +1243,143 @@ fun ContactRow(icon: ImageVector, text: String, textTag: String, buttonTag: Stri
             overflow = TextOverflow.Ellipsis,
             modifier = Modifier.weight(1f).testTag(textTag))
       }
+}
+
+/**
+ * Composable function representing the availability section.
+ *
+ * @param week List of opening hours for each day of the week.
+ * @param onEdit Callback function to handle editing of opening hours for a specific day.
+ */
+@Composable
+fun AvailabilitySection(week: List<OpeningHours>, onEdit: (Int) -> Unit) {
+    Column(Modifier.testTag(ShopFormTestTags.AVAILABILITY_LIST)) {
+        week.forEach { oh ->
+            val day = oh.day
+            DayRow(
+                dayName = ShopFormUi.dayNames[day], value = humanize(oh.hours), onEdit = { onEdit(day) })
+        }
+    }
+    Spacer(Modifier.height(Dimensions.Spacing.small))
+}
+
+/**
+ * Composable function representing a collapsible section with a title, optional header, and
+ * content.
+ *
+ * @param title The title of the section.
+ * @param initiallyExpanded Boolean indicating whether the section is initially expanded.
+ * @param header Optional composable function for the header content.
+ * @param content Composable function for the main content of the section.
+ * @param testTag Optional test tag for the section.
+ */
+@Composable
+fun CollapsibleSection(
+    title: String,
+    initiallyExpanded: Boolean = true,
+    header: (@Composable RowScope.() -> Unit)? = null,
+    content: @Composable ColumnScope.() -> Unit,
+    testTag: String? = null,
+    expanded: Boolean? = null,
+    onExpandedChange: ((Boolean) -> Unit)? = null,
+) {
+    val (isExpanded, setExpanded) =
+        if (expanded != null && onExpandedChange != null) {
+            expanded to onExpandedChange
+        } else {
+            var localExpanded by rememberSaveable { mutableStateOf(initiallyExpanded) }
+            localExpanded to { v: Boolean -> localExpanded = v }
+        }
+
+    val arrowRotation by
+    animateFloatAsState(
+        targetValue = if (isExpanded) Dimensions.Angles.expanded else Dimensions.Angles.collapsed,
+        label = "arrow")
+
+    Column(Modifier.fillMaxWidth()) {
+        Row(
+            modifier =
+                Modifier.fillMaxWidth().padding(top = Dimensions.Padding.medium).clickable { setExpanded(!isExpanded)}.testTag(testTag + ShopFormTestTags.SECTION_TOGGLE_SUFFIX),
+            verticalAlignment = Alignment.CenterVertically) {
+            Text(
+                text = title,
+                style = MaterialTheme.typography.titleMedium,
+                modifier =
+                    Modifier.weight(1f).let { m ->
+                        if (testTag != null) m.testTag(testTag + ShopFormTestTags.SECTION_TITLE_SUFFIX)
+                        else m
+                    })
+
+            header?.invoke(this)
+
+            IconButton(
+                onClick = { setExpanded(!isExpanded) },
+                modifier =
+                    Modifier.let { m ->
+                        if (testTag != null) m.testTag(testTag + ShopFormTestTags.SECTION_TOGGLE_SUFFIX)
+                        else m
+                    }) {
+                Icon(
+                    Icons.Filled.ExpandMore,
+                    contentDescription =
+                        if (isExpanded) ShopFormUi.Strings.COLLAPSE else ShopFormUi.Strings.EXPAND,
+                    modifier = Modifier.rotate(arrowRotation))
+            }
+        }
+
+        HorizontalDivider(
+            thickness = Dimensions.DividerThickness.standard,
+            color = MaterialTheme.colorScheme.outlineVariant,
+            modifier =
+                Modifier.padding(bottom = Dimensions.Spacing.large).let { m ->
+                    if (testTag != null) m.testTag(testTag + ShopFormTestTags.SECTION_DIVIDER_SUFFIX)
+                    else m
+                })
+
+        AnimatedVisibility(visible = isExpanded) {
+            Column(
+                Modifier.padding(top = Dimensions.Spacing.none).let { m ->
+                    if (testTag != null) m.testTag(testTag + ShopFormTestTags.SECTION_CONTENT_SUFFIX) else m
+                },
+                content = content)
+        }
+    }
+}
+
+/**
+ * Composable function representing the opening hours editor dialog.
+ *
+ * @param show Boolean indicating whether to show the dialog.
+ * @param day The day of the week being edited.
+ * @param week List of opening hours for each day of the week.
+ * @param onWeekChange Callback function to update the opening hours for the week.
+ * @param onDismiss Callback function to dismiss the dialog.
+ */
+@Composable
+fun OpeningHoursEditor(
+    show: Boolean,
+    day: Int?,
+    week: List<OpeningHours>,
+    onWeekChange: (List<OpeningHours>) -> Unit,
+    onDismiss: () -> Unit
+) {
+    if (!show || day == null) return
+    Box(Modifier.testTag(ShopFormTestTags.OPENING_HOURS_DIALOG_WRAPPER)) {
+        OpeningHoursDialog(
+            initialSelectedDays = setOf(day),
+            current = week[day],
+            onDismiss = onDismiss,
+            onSave = { selectedDays, closed, open24, intervals ->
+                val encoded: List<TimeSlot> =
+                    when {
+                        closed -> emptyList()
+                        open24 -> listOf(TimeSlot(TimeUi.OPEN24_START, TimeUi.OPEN24_END))
+                        else -> intervals.map { TimeSlot(it.first.hhmm(), it.second.hhmm()) }
+                    }
+                val copy = week.toMutableList()
+                selectedDays.forEach { d -> copy[d] = OpeningHours(day = d, hours = encoded) }
+                onWeekChange(copy)
+                onDismiss()
+            })
+    }
 }
